backend/requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
flask-swagger-ui==4.11.1
fonttools==4.54.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
six==1.16.0
tzdata==2024.2
Werkzeug==3.1.3
flasgger==0.9.7b2
flask_cors==5.0.0

